edit migrations
initial implementation of a script

+++ utils/regression-analysis/migrations/2022-05-13-112117_create_tests/up.sql
@@ -1,6 +1,6 @@
 -- Your SQL goes here
 CREATE TABLE tests (
-  id BIGINT PRIMARY KEY NOT NULL,
+  id INTEGER PRIMARY KEY NOT NULL,
   crate_name VARCHAR NOT NULL,
   test_name VARCHAR NOT NULL
 );
