+++ utils/regression-analysis/src/schema.rs
@@ -1,7 +1,7 @@
 table! {
     test_executions (id) {
-        id -> BigInt,
-        test_id -> BigInt,
+        id -> Integer,
+        test_id -> Integer,
         commit_hash -> Text,
         date_time -> Integer,
         exec_time -> BigInt,
@@ -10,7 +10,7 @@ table! {
 
 table! {
     tests (id) {
-        id -> BigInt,
+        id -> Integer,
         crate_name -> Text,
         test_name -> Text,
     }
